Made changes to code regarding api testing errors

+++ src/main/java/com/flightapp/flightbookingwebflux/model/Flight.java
@@ -1,8 +1,9 @@
 package com.flightapp.flightbookingwebflux.model;
 
-import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
+import lombok.Data;
+
 
 @Data
 @Document(collection = "flights")
@@ -12,20 +13,12 @@ public class Flight {
     private String id;
 
     private String airline;
-
     private String fromPlace;
-
     private String toPlace;
-
     private String flightName;
-
     private String departureTime;
-
     private String arrivalTime;
-
     private String flightDate;
-
     private int totalSeats;
-
-    private double price;
+    private int price;
 }
